+++ patzilla-ui/navigator/style/index.js
@@ -1,2 +1,10 @@
-require('./generic.css');
-require('./app.css');
+/*
+// -*- coding: utf-8 -*-
+// (c) 2018 Andreas Motl <andreas.motl@ip-tools.org>
+*/
+
+/* Material Components for the web */
+require('./material.css');
+
+/* Application styles */
+require('./index.css');